not all posts deleted because of null item fixed

+++ TP/Methods/actions/firebaseDB.cs
@@ -203,12 +203,15 @@ namespace TP.Methods.actions
                 var posts = JsonConvert.DeserializeObject<List<SubjectPosts>>(postData.Body);
                 foreach (var item in posts)
                 {
-                    if (item.SubId == subId)
+                    if(item != null)
                     {
-                        var postDelete = await client.DeleteAsync($"post/{item.PostId}");
-                        if (postDelete.StatusCode == System.Net.HttpStatusCode.OK)
+                        if (item.SubId == subId)
                         {
-                            deletedCount++;
+                            var postDelete = await client.DeleteAsync($"post/{item.PostId}");
+                            if (postDelete.StatusCode == System.Net.HttpStatusCode.OK)
+                            {
+                                deletedCount++;
+                            }
                         }
                     }
                 }
